refactor: change test folder for unique locale
